+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Added a `NEWS.md` file to track changes to the package.
 * Added fMRI data set: 'fMRI_localizer'.
+* Added Harvard-Orford brain atlas: 'brainAtlas'.
 
 # sansSouci.data 0.1.0